⚡️Compressing id and score values to reduce the number of bytes saved to datatstore. Assuming that the maximum length of allowed numbers are 12 digits, this was previously represented as 12 bytes. Now, a max of 8 bytes are needed for this number (1.5x saving). In some cases more bytes will be saved.
The previous save format was still much less efficient than this because it was stored as json, e.g. one entry would be `{"score":123456123456,"id":123456123456},`. This is now represented as 4 utf-8 characters - a reduction from 40 bytes to 16 bytes.

The records are only decompressed when they are needed. As binarySearch is used to locate the records, the number of decompressions required is O(log n). This has been implemented in leaderboardHelper.lua. Coding this was honestly a nightmare

✅ Added tests for leaderboardHelper.lua and utf8Compress.spec.lua